Updated Changed the name of 'web/main.html' to use 'index.html', which is the more traditional web file name.
Changed the alignment of things in index.html to be more human readable,
without impacting the code. Also chaged the name of the login prompt
from 'center' to 'loginPrompt', and aligned everything to center.

Change the font from using the systems monospace font, to the
OpenSans[1] font from the Google Fonts page, meaning that it will always
load this font on Win, Mac, Linux and F/N/O BSD.

==========> Links <==========
[1] - Open Sans on Google Fonts
https://fonts.google.com/specimen/Open+Sans